Issue #50: MediaPortal crashed after playback of a video in another plug-in, if and only if Moving Pictures had not been opened since Media Portal had been launched. This is fixed.
Issue #51: Focus on the "cycle view" button was lost when changing views. Because of this, the user would click to change views, click again to goto the next view but instead would be taken to the details screen since the facade now had focus. This is now fixed.

Issue 34: Cycling quickly through the facade caused massive slow downs because with every movement the plug-in would try to load fanart for the currently highlighted movie. I added a delay for switching artwork and immediately hid the current artwork when the user first clicks the button. The intention is to have the skinner create a fade animation. This issue needs more work though. The fade doesn't seem to be working properly and while performance is much better, there are still some jitters.