Issue #44: The GUI will now listen for messages indicating a movie has been added to or removed from the database and add or remove it from the facade as necessary. Also this issue uncovered a bug with the file system monitoring logic (more specifically on the database layer). This is also fixed.
Issue #28: Added basic removable media support in the GUI. If a file flagged as removable is missing it will prompt the user to insert the media. Still going to leave the issue open though, there is a lot that needs to be done to make this functionality more usable.

Minor tweaks in preperation for 0.5.3:
* Updated description for import paths on the config screen for clarity.
* Turned off filter button for now on the importer screen of the GUI as it is not yet implemented.