All changes in this check-in are from cyberfix73. There are further changes that I have confirmed resolve issue #43 (keeping track of the currently selected movie after playback or exiting and reentering the plug-in). I need to spend more time reviewing these changes though. Will get them checked in tomorrow afternoon. ----------------------------
Issue #88: Cover art added after clicking cancel
Solution: Now Gets the result of the dialog box and only updates if the user clicked OK

Issue #100: Some HD quality sample files squeeze through sample filter
Solution: Changed the default setting to 150. Will only be 150 on new installs by default. It does not update the old database setting. That can still be configured in the configuration options.

Issue #78: rating stars not working
Solution: If the value is over the maximum of 10.0, it will now be scaled down. Although, this really means that somebody put the wrong value there such as running time. At least it will work if on a different scale were used with a correct value. For example, a 679 will be converted to 6.8 just to it is on the right scale. This is only checked on importing because it would slow down movie scrolling a little to do it on the fly.

Issue #74: File names with underscores are not parsed properly
Solution: Now Replaces the underscores with periods like what is done with periods.