Issue 103: Partial implementation of the NFO scanner.
It will scan nfo for imdb and if found add the imdb id to the signature but it will not affect the matching yet.
Currently it only scans the mediamatches when folderhint is true and the scanner is not configurable yet. Will add this asap.
You can check the found id by clicking the rescan icon in the match overview.