Issue #87: Fully implemented local cover-art importing. Uses the same setup as the backdrops importer, anything with the name %imdb%.jpg, png, or bmp will get loaded that is in the cover-art folder. It will also load a folder.jpg (or png or bmp) file that is in the same folder as the movie files IF AND ONLY IF a corresponding setting is turned on under advanced settings. The default is to NOT load folder.jpg files.
I am going to leave the issue open because I want to add logic to load previously loaded cover art that is in the covers folder. If for example a user has rebuilt their database after movie-xml.com went down, but they did not delete their cover-art folder, then it would detect the covers in the cover-art folder that were previously loaded.