Issue #87: Added logic to load cover-art from previous runs of Moving Pictures. For example if the user rebuilt their database, but did not clear out their thumbs folder, then there will be some cover images previously retrieved from movie-xml.com. Reimporting these covers is not possible because movie-xml.com is down, but the local data provider will now automatically reload these anyway.
Issue #109: Redesigned some of the loading logic for the DataProviderManager. Was using reflection before to pull in all Types implementing IMovieProvider, but this really brings little benefit, and only complicates the code. New system requires rebuild for new default scanners but of course new scripts can be (or will be) loaded on the fly. More work needed on this setup of course.