Issue #67: Work for this issue done by vinteo. Thanks! The changes from vinteo's patch create additional strings piped to the GUI for fancy runtime formating.
Issue #123: Sometimes part two of a movie is played after an unrelated media file completes playback in another plug-in. I think this had to do with the autoplay feature of mountable images. If the user clicks play but on the "autoplay" menu, selects no, do not play, the currentlyPlaying flag is still set. I am not sure how to get the user's selection information from the autoplay dialog at this point in time, we will need to figure out how to do this eventually. For now I modified the code for when the user is exiting the plug-in, to automatically clear the currentlyPlaying flag. This will fix the issue, but could potentially cause problems for multi-part movies when playing in windowed mode while in another plug-in, so this is only a temporary solution.