Issue #142: Added additional checking to the GUI code to handle situations when the genre list is empty. Was causing an error when trying to index the first item in the list when the list had no members.
Issue #131: The database currently sometimes get's itself into a corrupted state for a very small subset of users. I have not been able to reproduce this issue yet, but I did make a change to prevent the program from crashing on startup. This should mean that the plug-in would still be usable, but one or two movies or local media records might have incorrect relation data in the DB. This could cause other errors. More work needs to be done on this issue.