Issue #145: Added a menu option from the main menu to toggle watched flag filtering. Still need some way to indicate to the user that they are looking at only Unwatched movies, or filtered movies.
Refactored signifigant portions of the MovingPicturesGUI class. Movie selection logic and facade management is all handled in the MovieBrowser class now. Also added a rudimentary and flexible filtering system for the facade/movie browser. Moved the Remote Numpad Filtering logic to a new filter using this class. Also used this system for the aforementioned watch flag filtering. There are some major changes to the code here so it will need to be tested thoroughly.

Temporarily disabled debug logging for scraper engine to more easily work with other aspects of the plug-in. Will remove the hard-cded value when Issue #138 is complete.