Issue #131: Alright I think I have fixed the slow shut down problems. In a recent version I added a background artwork scanner thread, that runs on startup. Once it scans all movies missing artwork it shuts itself down, but if the user shuts the plug-in down first, the process keeps running in the background until the artwork scanner completes. It is now shut down properly like everything else. This needs some 3rd party verification.
I also added some additional logic in a handful of classes to ensure that threads properly abort when told to. In some situations it might have been possible for the ThreadAbortException to get incorrectly handled in some vanilla exception handling.

I think the only delays now for when shutting down will be the timeout for web requests. because this is unmanaged code, I am not sure it is possible to force it to break until it returns to our logic...

In the future we should create a thread manager to handle this sort of stuff. There are various other threads created all throughout the config and the MP GUI. They all are minor, taking a few seconds and mainly exist to boost responsiveness, but still, these should all be tracked and probably shut down forcefully on application termination.