Issue #154: I think implementing a full unignore panel to bring back previously ignored movies is not totally needed for this release, so I just wen ahead and added a bulk unignore feature for the time being. It is not ideal and will be replaced later, but it is a messing fix for a situation that previously caused some people to have to rebuild their database.
Issue #117: I added a progress bar for long operations for updating movie details and added a more clear method to update movie details for all movies in the database.

Issue #109: Started the popup for the select script version functionality of the data source panel. When complete this will finalize the panel.