Issue #109: Finalized the data source panel. Selectable versions of scripts is now fully implemented and script language is properly displayed in the list (localized to the user's language).
A possible area of improvement is a published date for scripts. I am conflicted about this though, because I think script writers will be lazy and not update the field in the script file. I am not sure how to address this. I think an "added on" date would be a poor substitute. The point would be to let the user know how out of date their current scripts are...