Issue #135: Added the ability for the user to customize the file pattern used for loading cover-art from the movie folder. Default is still folder.jpg. Also added "old backdrop" so backdrops will be reloaded locally when the DB is rebuilt. Also added backdrop loading functionality from the movie folder. Like for covers this is off by default, but when turned on the default pattern is backdrop.jpg (user customizable).
It is worth noting that the code in the LocalProvider class has some redundancies. I am going to create a new Issue for cleaning up this class. It should be pretty easy to cut the number of lines in this class by 30-40%.

Issue #84: Artwork was not always "sticking" after loading. The artwork loading function were not setting an internal variable to flag that a commit was needed, so when the client (the art loading thread) called commit, the DB Manager figured the object didn't NEED to be commited, so it didn't commit. This issue should be fixed, now, I was able to reproduce and fix.