Issue #198: Previously the backdrop loader that pulls from the movie directory was failing because it was trying to index into the LocalMedia array before it was populated. This functionality would have only worked on movies that had already been loaded. This is now corrected.
Issue #197: Blank fields were being displayed as #MovingPictures.SelectedMovie.fieldname. Updated the parsing to populate it with a blank now and tweaked the image loading code to prevent new problems arising there (the GUITextureManager class doesn't check for a blank filename as invalid for some reason).

Issue #199: Sometimes files were not added to the importer. I think this could be because a DBLocalMedia object had been committed without it's attached DBMovieInfo. I think this would be a rare case, but when this happens it could cause the problem described by the user. Need to verify with the user if this fixed the problem. Also tweaked the logging and formatting in the path scanning logic of the DBImportPath class.