Issue #68: Image loading is now done asynchronously rather than by MediaPortal itself in the render thread. This should improve the responsiveness in the GUI, reducing the times when there is a delayed movement with a button press on the facade.
Issue #36: If the skin supports it, the plug-in will now smoothly fade from one backdrop to another. This change is backwards compatible. Old skins work fine with the new version of the plug-in (just no fading) and new skins work fine with the old version of the plug-in (again no fading). This does require a skin change though.