Follow-up on code review changes for r329:
- Moved DeviceMonitor out of configuration gui framework into MovieImporter (Still a initiator exist there to pass the winform handle into the DeviceMonitor until we come up with another generic monitor) 

Note: When the importer is started from within the normal mediaportal gui the listeners rely on the GUIMessage events. MediaPortal's AutoPlay listens to the same events so this will work regardless of AutoPlay being enabled. Still this is a MediaPortal dependancy we have to live with until we have our own monitor.

- Reverted changes to MovieBrowser class (the functionality it added was already present)