Issue #38: Filtering. Implemented advanced filtering classes, database changes required to support them and GUI required to edit them. This is a generic implementation so it should be suitable for filtering any type of database object (not just movies) should the need arise. Other plugins or applications based on Cornerstone should also be able to just drop in this filtering functionality.
NOTE: This is only phase one of filtering. It is not implemented into the GUI yet and basic filters and categories have not yet been implemented. Coming soon!

Issue #11: Linked an advanced filter to parental controls functionality. This also is not yet hooked into the GUI but GUI integration should be fairly trivial. One step at a time. 

**** See the bottom of the GUI Settings tab for new functionality regarding both parental controls and filtering.