Related to Issue #465: Implemented threaded initialization for GUI plugin
This will "dramatically" speed up the loading of mediaportal for users with a large collection. Before this change mediaportal would sit on "loading plugins" without the user knowing what's going on (when moving picture was checking deleted files etc). I changed it for moving pictures so that the initialization is performed in the background. Now when the user enters the plugin before this initialization is finished he will see a loading dialog with a progress bar much like the loading screen in the configuration gui. If the initialization was finished before the user enters the plugin the user won't notice anything. Thus the user won't be bothered by the initialization time of moving pictures when he starts MediaPortal and also by presenting the loading screen we update the user with progress on loading (instead of just the "loading plugins.." text on MediaPortal startup).

ps. If you have a small collection it's very likely that you won't notice any change at all.