Fixed: Renaming/moving files while the plugin was running could result in a movie being deleted and re-added.
the file system watcher now has a 'grace period' before it actually removes a file from the database. This will allow the importer to re-link a "new" file to it's removed counterpart (using the logic introduced in the previous revision). This also means we now support moving files/folders between import paths.