* Fixed bug preventing custom intros from playing unless resume data was completely absent (resume time = 0 seconds). Now uses the same limit as resume prompting (30 seconds). Possibly related to Issue #575.
Issue #662: Added logic to delete a recently saved image file if GDI failed saving the file and threw an exception. This in most cases means the image file is corrupt and often it is incorrectly flagged by virus scanners as an image file featuring an exploit. 

Issue #662: Also for this issue I changed the filetype for saving backdrops and covers back to PNG (and changed the extension to match). The GDI error referenced above seems to only occur when saving an image file as a JPG, so using the PNG file format results in more artwork available to the user. Also updated local cover art scanner to account for the new extension.