Issue #995: Improved file removal logic to prevent unneeded re-import (and loss) of existing movie entries.
For developers: all relevant local media removals are now handled by the importer. For this purpose a new public method was introduced called MarkMediaForRemoval() (method names, as always, are subject to change). The method will queue the local media object for removal in the existing internal queue. This queue has been improved to post-pone actual removal until all pending actions that might conflict with the removal are processed.