Issue #715: Adjusted article removal logic to quit if it found a valid language but there is no list of articles defined for that language. Best not to remove articles from another language if we know it doesn't match.
Also fixed a bug causing the article logic to fail. Was grabbing the full language string rather than the language code before.

Also moved the parsing logic for the article groups to MovingPicturesSettings. We provide helper properties like this for a few other settings as well when it makes sense.

Issue #1077: Fixed issue with change in r1077 that caused the new Title (Begins With) dynamic category to fail to use the Sort By field. Title will now be properly remapped to Sort By. For this to take effect for SVN testers you will need to rebuild the menu (remove everything).