Update filmtipset.se.xml (#1160)
* Update filmtipset.se.xml

Support for Release Date. This information is show in the Technical Details of the Titan skin. In the old version this field seems to instead show the date imported.
Changed data source URL from nyheter24.se/filmtipset to filmtipset.se.
If IMDb.com does not have a release date, then we will use the release year.

* Update filmtipset.se.xml

Support for the new filmtipset.se site.